move and rename MysteriumClient to mysterium-client
